+++ clic/src/tier1.cpp
@@ -1384,9 +1384,9 @@ range_func(const Device::Pointer & device,
            int                     stop_z,
            int                     step_z) -> Array::Pointer
 {
-  correct_range(&start_x, &stop_x, &step_x, src->width());
-  correct_range(&start_y, &stop_y, &step_y, src->height());
-  correct_range(&start_z, &stop_z, &step_z, src->depth());
+  correct_range(&start_x, &stop_x, &step_x, static_cast<int>(src->width()));
+  correct_range(&start_y, &stop_y, &step_y, static_cast<int>(src->height()));
+  correct_range(&start_z, &stop_z, &step_z, static_cast<int>(src->depth()));
   tier0::create_dst(src, dst, abs(start_x - stop_x), abs(start_y - stop_y), abs(start_z - stop_z), src->dtype());
   const KernelInfo    kernel = { "range", kernel::range };
   const ParameterList params = { { "src", src },         { "dst", dst },         { "start_x", start_x },
